requirements/dev.txt
-r base.txt

django-extensions==1.7.2
coverage==4.2
nose==1.3.7
model-mommy==1.2.6
django-nose==1.4.4
django-test-without-migrations==0.4